requirements.txt
discord.py
revChatGPT
